Relocated static third party library files, removed project files and isolated the package into a single AssemblyInfo
Instead of producing weird unstable undocumented / unaccounted for in-package dependencies we inline the namespaces as much as possible. This way there can't be any collision with consumer projects referencing other versions of CBOR or Chaos.NaCl.
I moved the libraries for the simple reason that isolating similar kinds of checked-in artifacts allows for better control. It's now possible to apply certain rules regarding third-party includes based on the "lib" directory.