Isolated test files into a single directory for a better solution explorer browsing experience
This is achieved by first moving the files into a subfolder and then update their existing ItemSpecs with the Link metadata.
This allows us to redirect the file output in the output directory. Using globbers we also removed the necessity to have a gazillion msbuild inclusion lines per test file.