Removed unnecessary UserSecretsID, rolled down test package references to release versions and removed the dotnet-xunit tool reference
The dotnet-sdk (based on the TestAdapter system) is perfectly capable of detecting, running and printing results for xunit projects.
As for the UserSecretsID, i removed this because none of the projects actually use the secrets system. I assume this was added accidentally - the current usersecret feature in the cli tends to spread into unrelated projects.