Add field userHandle in verifyAssertionWithServer
For the usernamaless result on the client side, the userHandle was missing in the reply. This can be nullable, but not missing since it is even backwards compatible according to: https://www.w3.org/TR/webauthn/#conforming-authenticators-u2f
however: BtB (by the book ;-)
Quote from: 
https://www.w3.org/TR/webauthn/#discover-from-external-source step: "->If any authenticator indicates success."
**userHandleResult**
If the authenticator returned a user handle, set the value of userHandleResult to be the bytes of the returned user handle. Otherwise, set the value of userHandleResult to null.
---