Intercede FIDO md2 new TOC root cert (#225)
The FIDO Alliance have recently changed the root certificate for their Metadata Service table of contents (TOC).  The certificates in the TOC certificate chain have also changed from being ECC to RSA.  This branch contains the new TOC root certificate and support for RSA in addition to ECC keys in the intermediate and signing certificates within the X5c data.